feat: manage reference images from the art style dialog
The dialog asked users to describe a style from reference images they
could neither see nor change without closing it and going back to the
assets section.

Extract the reference tiles, the add tile and the upload input into a
ReferenceImages component, and render it in both places. The dialog now
shows the same images the assets section does, and adding or removing
one there updates the project directly, since both read the store.

Also constrain the derive prompt to return only the style description,
so the model stops prefacing it with a sentence of its own.
