+++ app/components/canvas/elements/AssetsSection.tsx
@@ -1,22 +1,20 @@
 "use client";
 
 import { memo, useState } from "react";
-import { ImagePlus, UserPlus } from "@/components/ui/icon";
+import { UserPlus } from "@/components/ui/icon";
 import { useProject } from "@/lib/project/useProject";
-import { useImageUpload } from "@/lib/upload/useImageUpload";
 import { AddAssetTile } from "./AddAssetTile";
 import {
 	ArtStyleAssetTile,
 	CharacterAssetTiles,
 	NarratorAssetTile,
-	ReferenceAssetTiles,
 } from "./AssetTiles";
 import { useAssetEditDialogs } from "./character/useAssetEditDialogs";
 import { CollapsibleHeader } from "./CollapsibleHeader";
+import { ReferenceImages } from "./ReferenceImages";
 
 function AssetsSectionComponent() {
 	const hydrated = useProject((s) => s.hydrated);
-	const addReferenceImages = useProject((s) => s.addReferenceImages);
 	const [collapsed, setCollapsed] = useState(false);
 	const {
 		openCreateCharacter,
@@ -25,11 +23,6 @@ function AssetsSectionComponent() {
 		openArtStyle,
 		dialogs,
 	} = useAssetEditDialogs();
-
-	const { openPicker, uploading, inputElement } = useImageUpload({
-		multiple: true,
-		onUpload: addReferenceImages,
-	});
 
 	if (!hydrated) return null;
 
@@ -52,16 +45,7 @@ function AssetsSectionComponent() {
 						Icon={UserPlus}
 						onClick={openCreateCharacter}
 					/>
-					<ReferenceAssetTiles />
-					<AddAssetTile
-						label="Reference"
-						ariaLabel="Add reference image"
-						Icon={ImagePlus}
-						onClick={openPicker}
-						disabled={uploading}
-						busy={uploading}
-					/>
-					{inputElement}
+					<ReferenceImages />
 				</div>
 			)}
 			{dialogs}
