+++ app/components/canvas/elements/style/ArtStyleModal.tsx
@@ -23,7 +23,8 @@ import {
 } from "@/lib/project/deriveArtStyle";
 import { getProjectStore } from "@/lib/project/store";
 import { useProject } from "@/lib/project/useProject";
-import { TextAreaField } from "../character/fields";
+import { FieldLabel, TextAreaField } from "../character/fields";
+import { ReferenceImages } from "../ReferenceImages";
 import { ArtStylePresets } from "./ArtStylePresets";
 
 export function ArtStyleModal({
@@ -79,14 +80,19 @@ function ArtStyleDialogBody({ onClose }: { onClose: () => void }) {
 			</DialogHeader>
 
 			<div className="-mx-1 flex min-h-0 flex-1 flex-col gap-4 overflow-y-auto px-1">
-				<div className="flex flex-col gap-2">
-					<TextAreaField
-						label="Description"
-						value={style}
-						onChange={setStyle}
-						placeholder="Describe the look of every image, or paste a full image prompt"
-						rows={5}
-					/>
+				<TextAreaField
+					label="Description"
+					value={style}
+					onChange={setStyle}
+					placeholder="Describe the look of every image, or paste a full image prompt"
+					rows={5}
+				/>
+
+				<section aria-label="Reference images" className="flex flex-col gap-2">
+					<FieldLabel>Reference images</FieldLabel>
+					<div className="flex flex-wrap gap-2">
+						<ReferenceImages />
+					</div>
 					<Button
 						type="button"
 						variant="outline"
@@ -103,7 +109,7 @@ function ArtStyleDialogBody({ onClose }: { onClose: () => void }) {
 						{deriving && <Spinner className="text-current" />}
 						Use my reference images
 					</Button>
-				</div>
+				</section>
 
 				<ArtStylePresets value={style} onSelect={setStyle} />
 			</div>
